Fixed a DLL-related bug with disappearing menus

+++ fzui/src/fzui/color.hpp
@@ -1,11 +1,15 @@
 #pragma once
 
+// std
+#include <cstdint>
+
+// FZUI
 #include "fzui/core/core.hpp"
 
 namespace fz {
   struct FZ_API Color {
-    int r;
-    int g;
-    int b;
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
   };
 }
